requirements.txt
flake8==6.0.0
pre-commit==3.2.2
pylint==2.17.2
pyright==1.1.302
pytest==7.3.0
pytest-cov==4.0.0
pytest-html==3.2.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
flask==2.3.2
customtkinter==5.2.0